Extract common test generators and config
