+++ benchmark/test_conversion.py
@@ -11,90 +11,10 @@ measured operation is only the bound ``to_diagram``, ``to_hypergraph`` or
 ``to_map`` method.
 """
 
-import os
-import time
-
 import pytest
 
-from discopy import compact
-from discopy.symmetric import Box, Ty
-
-
-_FULL = "bench:full" in os.environ.get("BENCH_FLAGS", "").lower()
-
-
-def sizes(*base, full=()):
-    """Include the heavy tail only under ``BENCH_FLAGS=bench:full``."""
-    return list(base) + (list(full) if _FULL else [])
-
-
-def case(group):
-    """Measure CPU time with garbage collection disabled."""
-    return pytest.mark.benchmark(
-        group=group, timer=time.process_time, disable_gc=True)
-
-
-ROUNDS, WARMUP = 3, 1
-
-
-def repeated(op, box, n):
-    """Combine ``n`` copies of ``box`` with ``op`` by repeated doubling."""
-    if n == 1:
-        return box
-    half = repeated(op, box, n // 2)
-    result = op(half, half)
-    return op(result, box) if n % 2 else result
-
-
-def reverse_permutation(factory, dom):
-    """Build the reversal of ``dom`` using native swaps and composition."""
-    if len(dom) <= 1:
-        return factory.id(dom)
-    head = factory.swap(dom[:-1], dom[-1:])
-    tail = factory.id(dom[-1:]).tensor(
-        reverse_permutation(factory, dom[:-1]))
-    return head.then(tail)
-
-
-def source_box(representation):
-    """An atomic endomorphism embedded in ``representation``."""
-    box = Box("f", Ty("x"), Ty("x"))
-    if representation == "Hypergraph":
-        return box.to_hypergraph()
-    if representation == "CMap":
-        return box.to_map()
-    return box
-
-
-def series(representation, n):
-    """A depth-``n`` source morphism."""
-    box = source_box(representation)
-    return repeated(lambda f, g: f.then(g), box, n)
-
-
-def tensor(representation, n):
-    """A width-``n`` source morphism."""
-    box = source_box(representation)
-    return repeated(lambda f, g: f.tensor(g), box, n)
-
-
-def permutation(representation, n):
-    """A routing-heavy source morphism on ``n`` wires."""
-    box = source_box(representation)
-    return reverse_permutation(type(box), box.dom ** n)
-
-
-def snake(representation, n):
-    """A snake made of ``n`` zipping cups and caps."""
-    factory = {
-        "Diagram": compact.Diagram,
-        "Hypergraph": compact.Hypergraph,
-        "CMap": compact.CMap,
-    }[representation]
-    x = compact.Ty("x")
-    cups = repeated(lambda f, g: f.tensor(g), factory.cups(x, x.r), n)
-    caps = repeated(lambda f, g: f.tensor(g), factory.caps(x.r, x), n)
-    return factory.id(x).tensor(caps).then(cups.tensor(factory.id(x)))
+from benchmark.config import ROUNDS, WARMUP, case, sizes
+from benchmark.generators import reverse_permutation, series, snake, tensor
 
 
 # --- k-fold series ---------------------------------------------------------
@@ -202,7 +122,7 @@ def test_tensor_map_to_hypergraph(benchmark, n):
 @case("permutation (Diagram -> CMap)")
 @pytest.mark.parametrize("n", sizes(5, 10, 20, full=(50,)))
 def test_permutation_diagram_to_map(benchmark, n):
-    morphism = permutation("Diagram", n)
+    morphism = reverse_permutation("Diagram", n)
     benchmark.pedantic(
         morphism.to_map, rounds=ROUNDS, warmup_rounds=WARMUP)
 
@@ -210,7 +130,7 @@ def test_permutation_diagram_to_map(benchmark, n):
 @case("permutation (Diagram -> Hypergraph)")
 @pytest.mark.parametrize("n", sizes(5, 10, 20, full=(50,)))
 def test_permutation_diagram_to_hypergraph(benchmark, n):
-    morphism = permutation("Diagram", n)
+    morphism = reverse_permutation("Diagram", n)
     benchmark.pedantic(
         morphism.to_hypergraph, rounds=ROUNDS, warmup_rounds=WARMUP)
 
@@ -218,7 +138,7 @@ def test_permutation_diagram_to_hypergraph(benchmark, n):
 @case("permutation (Hypergraph -> Diagram)")
 @pytest.mark.parametrize("n", sizes(5, 10, 20, full=(50,)))
 def test_permutation_hypergraph_to_diagram(benchmark, n):
-    morphism = permutation("Hypergraph", n)
+    morphism = reverse_permutation("Hypergraph", n)
     benchmark.pedantic(
         morphism.to_diagram, rounds=ROUNDS, warmup_rounds=WARMUP)
 
@@ -226,7 +146,7 @@ def test_permutation_hypergraph_to_diagram(benchmark, n):
 @case("permutation (Hypergraph -> CMap)")
 @pytest.mark.parametrize("n", sizes(5, 10, 20, full=(50,)))
 def test_permutation_hypergraph_to_map(benchmark, n):
-    morphism = permutation("Hypergraph", n)
+    morphism = reverse_permutation("Hypergraph", n)
     benchmark.pedantic(
         morphism.to_map, rounds=ROUNDS, warmup_rounds=WARMUP)
 
@@ -234,7 +154,7 @@ def test_permutation_hypergraph_to_map(benchmark, n):
 @case("permutation (CMap -> Diagram)")
 @pytest.mark.parametrize("n", sizes(5, 10, 20, full=(50,)))
 def test_permutation_map_to_diagram(benchmark, n):
-    morphism = permutation("CMap", n)
+    morphism = reverse_permutation("CMap", n)
     benchmark.pedantic(
         morphism.to_diagram, rounds=ROUNDS, warmup_rounds=WARMUP)
 
@@ -242,7 +162,7 @@ def test_permutation_map_to_diagram(benchmark, n):
 @case("permutation (CMap -> Hypergraph)")
 @pytest.mark.parametrize("n", sizes(5, 10, 20, full=(50,)))
 def test_permutation_map_to_hypergraph(benchmark, n):
-    morphism = permutation("CMap", n)
+    morphism = reverse_permutation("CMap", n)
     benchmark.pedantic(
         morphism.to_hypergraph, rounds=ROUNDS, warmup_rounds=WARMUP)
